# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/faculties": {
      "post": {
        "description": "Creates a new faculty record.",
        "operationId": "FacultiesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFacultyInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Faculty created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Faculty"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Create a new faculty",
        "tags": [
          "faculties"
        ]
      },
      "get": {
        "description": "Returns a list of all faculties.",
        "operationId": "FacultiesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "A list of faculties.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Faculty"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Retrieve all faculties",
        "tags": [
          "faculties"
        ]
      }
    },
    "/faculties/bulk": {
      "post": {
        "description": "Creates multiple faculty records in a single operation.",
        "operationId": "FacultiesController_createMany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManyFacultiesInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The number of faculties created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "number",
                      "description": "Number of faculties created",
                      "example": 5
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Create multiple faculties",
        "tags": [
          "faculties"
        ]
      }
    },
    "/faculties/{id}": {
      "get": {
        "description": "Returns a single faculty by its ID.",
        "operationId": "FacultiesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the faculty",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The faculty with the specified ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Faculty"
                }
              }
            }
          },
          "404": {
            "description": "Faculty not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Retrieve a faculty by ID",
        "tags": [
          "faculties"
        ]
      },
      "put": {
        "description": "Updates an existing faculty record.",
        "operationId": "FacultiesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the faculty",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The updated faculty.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Faculty"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "404": {
            "description": "Faculty not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Update a faculty by ID",
        "tags": [
          "faculties"
        ]
      },
      "delete": {
        "description": "Deletes a faculty record by its ID.",
        "operationId": "FacultiesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the faculty",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Faculty deleted"
          },
          "404": {
            "description": "Faculty not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Delete a faculty by ID",
        "tags": [
          "faculties"
        ]
      }
    },
    "/faculties/university/{universityId}": {
      "get": {
        "description": "Returns a list of faculties associated with a specific university ID.",
        "operationId": "FacultiesController_findByUniversity",
        "parameters": [
          {
            "name": "universityId",
            "required": true,
            "in": "path",
            "description": "The ID of the university",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of faculties for the specified university.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Faculty"
                  }
                }
              }
            }
          },
          "404": {
            "description": "University not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Retrieve faculties by university ID",
        "tags": [
          "faculties"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lumina API",
    "description": "API documentation for Lumina server",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "Translation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "en_text": {
            "type": "string"
          },
          "he_text": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "en_text",
          "he_text"
        ]
      },
      "Faculty": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "description": {
            "$ref": "#/components/schemas/Translation"
          },
          "degrees": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Degree"
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "degrees"
        ]
      },
      "Degree": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "universityId": {
            "type": "string"
          },
          "university": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/University"
              }
            ]
          },
          "courses": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Course"
            }
          },
          "facultyId": {
            "type": "string",
            "nullable": true
          },
          "faculty": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Faculty"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "universityId",
          "university",
          "courses",
          "facultyId",
          "faculty"
        ]
      },
      "University": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "courses": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Course"
            }
          },
          "degrees": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Degree"
            }
          }
        },
        "required": [
          "id",
          "name",
          "courses",
          "degrees"
        ]
      },
      "RelationshipMetadataKey": {
        "type": "string",
        "enum": [
          "REASON",
          "TYPE",
          "DESCRIPTION"
        ]
      },
      "RelationshipMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "key": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RelationshipMetadataKey"
              }
            ]
          },
          "value": {
            "type": "string"
          },
          "blockRelationshipId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "key",
          "value",
          "blockRelationshipId"
        ]
      },
      "BlockRelationship": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "prerequisiteId": {
            "type": "string"
          },
          "postrequisiteId": {
            "type": "string"
          },
          "prerequisite": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Block"
              }
            ]
          },
          "postrequisite": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Block"
              }
            ]
          },
          "metadata": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelationshipMetadata"
            }
          }
        },
        "required": [
          "id",
          "prerequisiteId",
          "postrequisiteId",
          "prerequisite",
          "postrequisite",
          "metadata"
        ]
      },
      "Block": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "prerequisiteFor": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockRelationship"
            }
          },
          "postrequisiteOf": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockRelationship"
            }
          },
          "modules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          }
        },
        "required": [
          "id",
          "prerequisiteFor",
          "postrequisiteOf",
          "modules"
        ]
      },
      "QuestionValidationStatus": {
        "type": "string",
        "enum": [
          "ai_generated",
          "in_manual_review",
          "approved",
          "rejected"
        ]
      },
      "SelectAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "translationId": {
            "type": "string"
          },
          "text": {
            "$ref": "#/components/schemas/Translation"
          },
          "answerId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "isCorrect",
          "translationId",
          "text",
          "answerId"
        ]
      },
      "UnitAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "answerId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "value",
          "unit",
          "answerId"
        ]
      },
      "NumberAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "answerId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "value",
          "answerId"
        ]
      },
      "Answer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "questionId": {
            "type": "string"
          },
          "SelectAnswer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SelectAnswer"
            }
          },
          "UnitAnswer": {
            "$ref": "#/components/schemas/UnitAnswer"
          },
          "NumberAnswer": {
            "$ref": "#/components/schemas/NumberAnswer"
          }
        },
        "required": [
          "id",
          "questionId"
        ]
      },
      "QuestionType": {
        "type": "string",
        "enum": [
          "selection",
          "value",
          "void"
        ]
      },
      "QuestionPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "questionId": {
            "type": "string"
          },
          "partQuestionId": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "questionId",
          "partQuestionId",
          "order"
        ]
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "validationStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionValidationStatus"
              }
            ]
          },
          "translationId": {
            "type": "string"
          },
          "text": {
            "$ref": "#/components/schemas/Translation"
          },
          "Answer": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Answer"
            }
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionType"
              }
            ]
          },
          "Modules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          },
          "Parts": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionPart"
            }
          },
          "PartOf": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionPart"
            }
          }
        },
        "required": [
          "id",
          "validationStatus",
          "translationId",
          "text",
          "Answer",
          "type",
          "Modules",
          "Parts",
          "PartOf"
        ]
      },
      "Module": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "translationId": {
            "type": "string"
          },
          "subModules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          },
          "parentModules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          },
          "Block": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Block"
              }
            ]
          },
          "questions": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            }
          }
        },
        "required": [
          "id",
          "name",
          "translationId",
          "subModules",
          "parentModules",
          "Block",
          "questions"
        ]
      },
      "Course": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "universityId": {
            "type": "string"
          },
          "university": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/University"
              }
            ]
          },
          "modules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          },
          "publishedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "Block": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Block"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "universityId",
          "university",
          "modules",
          "publishedAt",
          "Block"
        ]
      },
...
      "CreateFacultyInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the faculty"
          },
          "description": {
            "type": "string",
            "description": "Description of the faculty"
          },
          "universityId": {
            "type": "string",
            "description": "ID of the university this faculty belongs to"
          }
        },
        "required": [
          "name",
          "description",
          "universityId"
        ]
      },
      "CreateManyFacultiesInput": {
        "type": "object",
        "properties": {
          "faculties": {
            "description": "Array of faculty data to create",
            "example": [
              {
                "name": "Computer Science",
                "description": "CS Department",
                "universityId": "uuid1"
              },
              {
                "name": "Mathematics",
                "description": "Math Department",
                "universityId": "uuid2"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateFacultyInput"
            }
          }
        },
        "required": [
          "faculties"
        ]
      },
...
    }
  }
}